requirements.txt
openai>=1.0.0
GitPython>=3.1.30
pytest>=7.0
